requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.26
openai==1.9.0
regex==2023.12.25
requests==2.31.0
tiktoken==0.5.2
psycopg2-binary==2.9.9
